directory 중 2

+++ lovemyself/src/main/java/com/psc/lovemyself/repository/findmyself/DirectoryRepository.java
@@ -3,5 +3,5 @@ package com.psc.lovemyself.repository.findmyself;
 import com.psc.lovemyself.domain.directory.Directory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DirectoryRepository extends JpaRepository<Directory, Integer> {
+public interface DirectoryRepository extends JpaRepository<Directory, Long> {
 }
